Integrate: [the Devel::PPPort parts of]
[ 30728]
Upgrade to Devel::PPPort 3.11_01

[ 31134]
move PL_expect and PL_copline into the PL_parser structure

[ 31157]
Make ppptools.pl allow a trailing dot in filenames for VMS.

[ 31199]
move PL_rsfp into the PL_parser struct
and simplify its creation and destruction

[ 31200]
move PL_rsfp_filters into the parser struct

[ 31705]
Upgrade to Devel::PPPort 3.11_02

[ 31713]
Upgrade to Devel::PPPort 3.11_03

[ 31739]
Upgrade to Devel::PPPort 3.11_05

[ 31848]
Upgrade to Devel::PPPort 3.11_06

[ 31939]
Upgrade to Devel::PPPort 3.12

[ 32023]
Upgrade to Devel::PPPort 3.13
p4raw-link: @32023 on //depot/perl: c58e738a0202cd01343a506a6f46375f9382b043
p4raw-link: @31939 on //depot/perl: e47d7ad1097fe8bf25456f76aaa863c85118bca8
p4raw-link: @31848 on //depot/perl: aab9a3b697e77c80c3ac8b3db60730d246a5fbf5
p4raw-link: @31739 on //depot/perl: c83e6f195f905dd4809cef6ea71ef6cef8c9f7b8
p4raw-link: @31713 on //depot/perl: af36fda705dfd5f8d8d5a60ea777cd672a80ccf3
p4raw-link: @31705 on //depot/perl: 679ad62ddea1877bd96eafca271767acf6241fd7
p4raw-link: @31200 on //depot/perl: 5486870fe7f0fd0e99bf9619d5fd857a5b972014
p4raw-link: @31199 on //depot/perl: 2f9285f84584cb56950bf07de6ded6ebcdc3d302
p4raw-link: @31157 on //depot/perl: 0f06b1bbd2d65b2f053b58aeebed39ac8aba7fc6
p4raw-link: @31134 on //depot/perl: 53a7735b62aee14640bc8ca619b4267f07d686b9
p4raw-link: @30728 on //depot/perl: a89b7ab83a4c77596899b9598311b5ab588d5f11

p4raw-id: //depot/maint-5.8/perl@32292
p4raw-branched: from //depot/perl@32291 'branch in'
	ext/Devel/PPPort/parts/ppport.fnc
	ext/Devel/PPPort/t/shared_pv.t
	ext/Devel/PPPort/parts/inc/shared_pv (@31739..)
p4raw-integrated: from //depot/perl@32291 'copy in'
	ext/Devel/PPPort/MANIFEST.SKIP (@23912..)
	ext/Devel/PPPort/devel/mkapidoc.sh
	ext/Devel/PPPort/devel/scanprov (@26866..)
	ext/Devel/PPPort/devel/devtools.pl
	ext/Devel/PPPort/devel/mktodo ext/Devel/PPPort/devel/mktodo.pl
	ext/Devel/PPPort/devel/regenerate
	ext/Devel/PPPort/parts/base/5004000
	ext/Devel/PPPort/parts/base/5006000
	ext/Devel/PPPort/parts/base/5007001
	ext/Devel/PPPort/parts/base/5008001
	ext/Devel/PPPort/parts/base/5009000
	ext/Devel/PPPort/parts/base/5009002
	ext/Devel/PPPort/parts/todo/5004000
	ext/Devel/PPPort/parts/todo/5007002
	ext/Devel/PPPort/parts/todo/5008001
	ext/Devel/PPPort/parts/todo/5009000 (@28332..)
	ext/Devel/PPPort/parts/apidoc.fnc (@28507..)
	ext/Devel/PPPort/devel/buildperl.pl
	ext/Devel/PPPort/parts/base/5005000
	ext/Devel/PPPort/parts/base/5007002
	ext/Devel/PPPort/parts/base/5009004
	ext/Devel/PPPort/parts/embed.fnc ext/Devel/PPPort/t/SvPV.t
	ext/Devel/PPPort/t/magic.t (@29436..)
	ext/Devel/PPPort/parts/base/5009005
	ext/Devel/PPPort/parts/todo/5009005 (@29437..)
	ext/Devel/PPPort/parts/inc/newCONSTSUB
	ext/Devel/PPPort/parts/inc/ppphdoc
	ext/Devel/PPPort/parts/inc/threads
	ext/Devel/PPPort/parts/inc/uv
	ext/Devel/PPPort/parts/inc/version
	ext/Devel/PPPort/parts/inc/warn (@30280..)
	ext/Devel/PPPort/t/variables.t (@30728..)
	ext/Devel/PPPort/parts/apicheck.pl
	ext/Devel/PPPort/parts/base/5009003
	ext/Devel/PPPort/parts/inc/call ext/Devel/PPPort/parts/inc/grok
	ext/Devel/PPPort/parts/inc/misc
	ext/Devel/PPPort/parts/inc/sv_xpvf (@31705..)
	ext/Devel/PPPort/parts/ppptools.pl (@31713..)
	ext/Devel/PPPort/parts/inc/magic (@31739..)
	ext/Devel/PPPort/parts/inc/variables (@31848..)
	ext/Devel/PPPort/Changes (@31939..)
p4raw-integrated: from //depot/perl@31739 'edit in' MANIFEST (@31732..)
	'ignore' ext/Devel/PPPort/TODO ext/Devel/PPPort/t/ppphtest.t
	(@31705..) ext/Devel/PPPort/parts/inc/ppphbin (@31713..)
p4raw-integrated: from //depot/perl@31713 'ignore'
	ext/Devel/PPPort/parts/inc/SvPV ext/Devel/PPPort/soak
	(@31705..)
p4raw-integrated: from //depot/perl@31705 'ignore'
	ext/Devel/PPPort/parts/todo/5007001 (@28332..)
	ext/Devel/PPPort/HACKERS ext/Devel/PPPort/parts/inc/ppphtest
	ext/Devel/PPPort/parts/inc/pvs (@30280..)
	ext/Devel/PPPort/PPPort_pm.PL (@30728..)
p4raw-integrated: from //depot/perl@30728 'ignore'
	ext/Devel/PPPort/parts/todo/5006000 (@28332..)
	ext/Devel/PPPort/t/call.t (@29436..)

+++ ext/Devel/PPPort/devel/buildperl.pl
@@ -5,13 +5,13 @@
 #
 ################################################################################
 #
-#  $Revision: 10 $
+#  $Revision: 12 $
 #  $Author: mhx $
-#  $Date: 2006/12/02 09:58:34 +0100 $
+#  $Date: 2007/08/12 15:06:31 +0200 $
 #
 ################################################################################
 #
-#  Version 3.x, Copyright (C) 2004-2006, Marcus Holland-Moritz.
+#  Version 3.x, Copyright (C) 2004-2007, Marcus Holland-Moritz.
 #  Version 2.x, Copyright (C) 2001, Paul Marquess.
 #  Version 1.x, Copyright (C) 1999, Kenneth Albanowski.
 #
@@ -31,6 +31,12 @@ use Archive::Tar;
 use Cwd;
 
 # TODO: - extra arguments to Configure
+
+#
+#  --test-archives=1      check if archives can be read
+#  --test-archives=2      like 1, but also extract archives
+#  --test-archives=3      like 2, but also apply patches
+#
 
 my %opt = (
   prefix  => '/tmp/perl/install/<config>/<perl>',
@@ -92,10 +98,37 @@ my @patch = (
   },
   {
     perl => [
-              qr/^5\.004_0[1234]/,
+              qr/^5\.004_0[1234]$/,
             ],
     subs => [
               [ \&patch_doio ],
+            ],
+  },
+  {
+    perl => [
+              qw/
+                5.005
+                5.005_01
+                5.005_02
+              /,
+            ],
+    subs => [
+              [ \&patch_sysv, old_format => 1 ],
+            ],
+  },
+  {
+    perl => [
+              qw/
+                5.005_03
+                5.005_04
+              /,
+              qr/^5\.6\.[0-2]$/,
+              qr/^5\.7\.[0-3]$/,
+              qr/^5\.8\.[0-8]$/,
+              qr/^5\.9\.[0-5]$/
+            ],
+    subs => [
+              [ \&patch_sysv ],
             ],
   },
 );
@@ -111,7 +144,7 @@ GetOptions(\%opt, qw(
   force
   test
   install!
-  test-archives+
+  test-archives=i
 )) or pod2usage(2);
 
 if (exists $opt{config}) {
@@ -149,6 +182,12 @@ if ($opt{'test-archives'}) {
   for my $perl (@perls) {
     eval {
       my $d = extract_source($perl{$perl});
+      if ($opt{'test-archives'} > 2) {
+        my $cwd2 = cwd;
+        chdir $d or die "chdir $d: $!\n";
+        patch_source($perl{$perl}{version});
+        chdir $cwd2 or die "chdir $cwd2:$!\n"
+      }
       rmtree($d) if -e $d;
     };
     warn $@ if $@;
@@ -296,13 +335,13 @@ sub build_and_install
 sub patch_db
 {
   my $ver = shift;
-  print "patching DB_File\n";
+  print "patching ext/DB_File/DB_File.xs\n";
   run_or_die("sed -i -e 's/<db.h>/<db$ver\\/db.h>/' ext/DB_File/DB_File.xs");
 }
 
 sub patch_doio
 {
-  patch('doio.c', <<'END');
+  patch(<<'END');
 --- doio.c.org	2004-06-07 23:14:45.000000000 +0200
 +++ doio.c	2003-11-04 08:03:03.000000000 +0100
 @@ -75,6 +75,16 @@
@@ -325,11 +364,52 @@ sub patch_doio
 END
 }
 
+sub patch_sysv
+{
+  my %opt = @_;
+
+  # check if patching is required
+  return if $^O ne 'linux' or -f '/usr/include/asm/page.h';
+
+  if ($opt{old_format}) {
+    patch(<<'END');
+--- ext/IPC/SysV/SysV.xs.org	1998-07-20 10:20:07.000000000 +0200
++++ ext/IPC/SysV/SysV.xs	2007-08-12 10:51:06.000000000 +0200
+@@ -3,9 +3,6 @@
+ #include "XSUB.h"
+ 
+ #include <sys/types.h>
+-#ifdef __linux__
+-#include <asm/page.h>
+-#endif
+ #if defined(HAS_MSG) || defined(HAS_SEM) || defined(HAS_SHM)
+ #include <sys/ipc.h>
+ #ifdef HAS_MSG
+END
+  }
+  else {
+    patch(<<'END');
+--- ext/IPC/SysV/SysV.xs.org	2007-08-11 00:12:46.000000000 +0200
++++ ext/IPC/SysV/SysV.xs	2007-08-11 00:10:51.000000000 +0200
+@@ -3,9 +3,6 @@
+ #include "XSUB.h"
+ 
+ #include <sys/types.h>
+-#ifdef __linux__
+-#   include <asm/page.h>
+-#endif
+ #if defined(HAS_MSG) || defined(HAS_SEM) || defined(HAS_SHM)
+ #ifndef HAS_SEM
+ #   include <sys/ipc.h>
+END
+  }
+}
+
 sub patch
 {
-  my($file, $patch) = @_;
-  print "patching $file\n";
-  my $diff = "$file.diff";
+  my($patch) = @_;
+  print "patching $_\n" for $patch =~ /^\+{3}\s+(\S+)/gm;
+  my $diff = 'tmp.diff';
   write_or_die($diff, $patch);
   run_or_die("patch -s -p0 <$diff");
   unlink $diff or die "unlink $diff: $!\n";
@@ -418,7 +498,7 @@ and don't install them, run:
 
 =head1 COPYRIGHT
 
-Copyright (c) 2004-2006, Marcus Holland-Moritz.
+Copyright (c) 2004-2007, Marcus Holland-Moritz.
 
 This program is free software; you can redistribute it and/or
 modify it under the same terms as Perl itself.
